add screens folder and arrow size fix
